Add Fastly Compute@Edge project.
This commit adds a project for the wasmbuilder.app website itself, which is
hosted on Fastly's Compute@Edge service as a static site.

Added a `fastly-static-site` crate for a proc-macro to create the C@E app from
the contents of the `dist` directory (created from `make bundle`).

There's no automation yet, but it wouldn't be hard to setup a publishing CI
that pushes the site to C@E.