Replace WebAssembly module with a component.
This commit replaces the graph WebAssembly module (and wasm-bindgen) with a
WebAssembly component (with wit-bindgen).

The component code gets a well-defined interface and doesn't have to use serde
to marshal the complex types to the JavaScript.

It also allows the JavaScript to directly use the generated bindings rather
than having to define the types on both sides.