Update dependencies to latest.
This commit updates all dependencies to latest.

As a result of updating to the most recent `wasm-compose`, the graph component
now uses a resource instead of exporting flat functions so that the graph state
can be tied to the lifetime of the resource.

This necessitated creating a Graph singleton where state is maintained in the
react app.